# Conflicts:
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_gui/choices.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_gui/fieldopt.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_gui/ndisplay.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_gui/plusabsd.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_gui/plusabsi.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/addlink1d2d.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/addlink1d2dinternal.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/addlink2d.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/fixedweirfriction2d.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/furu_structures.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/getcz.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/getczz0.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/getprof2d.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/getprof_1d.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/getseg1d.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/getustbcfuhi.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/hydraulicallysmooth.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/iterfurufm.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/setbaptist.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/setcfuhi.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/setumod.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/yzprofile.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute_sediment/fm_bedform.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/model_specific/belanger.f90

+++ src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/prepost/updatevaluesonlaterals.f90
@@ -35,17 +35,17 @@
 subroutine updateValuesOnLaterals(tim1, timestep)
    use m_flowtimes, only: ti_his, time_his, ti_hiss
    use m_laterals, only: qqLat, numlatsg, num_layers, qplat, qplatCum, qplatCumPre, qplatAve, qLatReal, &
-                        qLatRealCum, qLatRealCumPre, qLatRealAve, n1latsg, n1latsg, n2latsg, nnlat
+                         qLatRealCum, qLatRealCumPre, qLatRealAve, n1latsg, n1latsg, n2latsg, nnlat
    use precision
    use m_alloc
    use m_flowparameters, only: eps10
    use m_partitioninfo, only: jampi, reduce_double_sum, is_ghost_node
    implicit none
-   double precision, intent(in) :: tim1 !< Current (new) time
-   double precision, intent(in) :: timestep !< Timestep is the difference between tim1 and the last update time
+   real(kind=dp), intent(in) :: tim1 !< Current (new) time
+   real(kind=dp), intent(in) :: timestep !< Timestep is the difference between tim1 and the last update time
 
    integer :: k1, i, i_lat, i_layer, i_node
-   double precision, allocatable :: qLatRealCumTmp(:), qLatRealMPI(:)
+   real(kind=dp), allocatable :: qLatRealCumTmp(:), qLatRealMPI(:)
 
    ! If current time has not reached the history output start time yet, do not update
    if (comparereal(tim1, ti_hiss, eps10) < 0) then
